# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TrustLoopGuard API
  description: Real-time guardrail runtime for AI agents.
  license:
    name: Apache-2.0
  version: 0.0.1
paths:
...
  /v1/financial/policies:
    get:
      tags:
      - financial
      operationId: list_policies
      responses:
        '200':
          description: Financial spending controls
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialPolicyListResponse'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - financial
      operationId: create_policy
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFinancialPolicyRequest'
        required: true
      responses:
        '201':
          description: Financial spending control created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialPolicyRecord'
        '400':
          description: Malformed or invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /v1/policies:
    get:
      tags:
      - policies
      summary: '`GET /v1/policies` — list active policy summaries.'
      operationId: list_policies
      responses:
        '200':
          description: All active policies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyListResponse'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - policies
      summary: '`POST /v1/policies` — create or update a policy from YAML or JSON.'
      operationId: upsert_policy
      requestBody:
        description: Policy document, YAML or JSON
        content:
          application/yaml:
            schema:
              type: string
        required: true
      responses:
        '201':
          description: Policy created or updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyDocument'
        '400':
          description: Malformed request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Policy failed validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/policies/batch/enabled:
    patch:
      tags:
      - policies
      summary: '`PATCH /v1/policies/batch/enabled` — enable or disable multiple policies.'
      operationId: batch_set_policy_enabled
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyBatchSetEnabledRequest'
        required: true
      responses:
        '200':
          description: Updated policies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyBatchSetEnabledResponse'
        '400':
          description: Malformed request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: One or more policies were not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /v1/policies/validate:
    post:
      tags:
      - policies
      summary: '`POST /v1/policies/validate` — validate policy YAML or JSON without saving it.'
      operationId: validate_policy
      requestBody:
        description: Policy document, YAML or JSON
        content:
          application/yaml:
            schema:
              type: string
        required: true
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyValidateResponse'
        '400':
          description: Malformed request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/policies/{id}:
    get:
      tags:
      - policies
      summary: '`GET /v1/policies/:id` — fetch a policy document.'
      operationId: get_policy
      parameters:
      - name: id
        in: path
        description: Policy identifier
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Policy found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyDocument'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Policy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags:
      - policies
      summary: '`DELETE /v1/policies/:id` — soft-delete a policy.'
      operationId: delete_policy
      parameters:
      - name: id
        in: path
        description: Policy identifier
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Policy deleted
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Policy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/policies/{id}/enabled:
    patch:
      tags:
      - policies
      summary: '`PATCH /v1/policies/:id/enabled` — enable or disable a policy.'
      operationId: set_policy_enabled
      parameters:
      - name: id
        in: path
        description: Policy identifier
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicySetEnabledRequest'
        required: true
      responses:
        '200':
          description: Updated policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyDocument'
        '400':
          description: Malformed request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Policy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
...
    ApiError:
      type: object
      description: |-
        Canonical error envelope returned on non-2xx responses from every
        TrustLoopGuard endpoint. SDKs deserialize this body to produce typed
        errors; integrators don't have to inspect status codes by hand.

        The shape is intentionally minimal — `code` drives SDK-side fan-out,
        `message` is for logs, `retriable` tells callers whether the same
        request may be retried, and `details` is opaque so the server can
        add validation field paths without breaking SDKs.
      required:
      - code
      - message
      - retriable
      properties:
        code:
          $ref: '#/components/schemas/ApiErrorCode'
        details:
          description: |-
            Opaque structured details (e.g. validation field path).
            Defaults to `null`; servers may add fields without breaking SDKs.
        message:
          type: string
        retriable:
          type: boolean
          description: |-
            Whether the caller may retry the same request without modification.
            SDKs honor `Retry-After` when present in addition to this flag.
    ApiErrorCode:
      type: string
      description: |-
        Stable error code dictionary. Add variants here when introducing new
        failure modes; never repurpose an existing variant — SDK callers may
        be branching on it.
      enum:
      - invalid
      - unauthorized
      - forbidden
      - not_found
      - gone
      - unprocessable
      - rate_limited
      - internal
      - unavailable
...
    CreateFinancialPolicyRequest:
      type: object
      required:
      - id
      - when
      properties:
        allowed_counterparty_ids:
          type: array
          items:
            type: string
        approval_threshold_minor:
          type:
          - integer
          - 'null'
          format: int64
        approver_roles:
          type: array
          items:
            type: string
        daily_minor:
          type:
          - integer
          - 'null'
          format: int64
        denied_counterparty_ids:
          type: array
          items:
            type: string
        description:
          type:
          - string
          - 'null'
        failed_precondition_action:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/PolicyAction'
        hold_above_minor:
          type:
          - integer
          - 'null'
          format: int64
        hold_new_counterparty:
          type: boolean
        id:
          type: string
        mandate_required:
          type: boolean
        missing_evidence_action:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/PolicyAction'
        monthly_minor:
          type:
          - integer
          - 'null'
          format: int64
        on_breach:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/PolicyAction'
        per_transaction_minor:
          type:
          - integer
          - 'null'
          format: int64
        refund_original_method_only:
          type: boolean
        required_preconditions:
          type: array
          items:
            $ref: '#/components/schemas/FinancialActionPrecondition'
        severity:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/Severity'
        when:
          $ref: '#/components/schemas/FinancialPolicySelector'
...
    FinancialActionKind:
      type: string
      enum:
      - payment
      - refund
      - payout
      - invoice_approval
      - purchase
      - treasury_transfer
      - consent
      - other
...
    FinancialActionPrecondition:
      type: string
      enum:
      - order_exists
      - payment_captured
      - refund_window_open
      - amount_lte_refundable_balance
      - destination_is_original_payment_method
      - no_duplicate_refund
      - invoice_matches_po
      - vendor_approved
      - mandate_valid
      - custom
...
    FinancialPolicyListResponse:
      type: object
      required:
      - policies
      properties:
        policies:
          type: array
          items:
            $ref: '#/components/schemas/FinancialPolicyRecord'
    FinancialPolicyRecord:
      type: object
      required:
      - id
      - severity
      - when
      - missing_evidence_action
      - failed_precondition_action
      - on_breach
      - enabled
      properties:
        allowed_counterparty_ids:
          type: array
          items:
            type: string
        approval_threshold_minor:
          type:
          - integer
          - 'null'
          format: int64
        approver_roles:
          type: array
          items:
            type: string
        daily_minor:
          type:
          - integer
          - 'null'
          format: int64
        denied_counterparty_ids:
          type: array
          items:
            type: string
        description:
          type:
          - string
          - 'null'
        enabled:
          type: boolean
        failed_precondition_action:
          $ref: '#/components/schemas/PolicyAction'
        hold_above_minor:
          type:
          - integer
          - 'null'
          format: int64
        hold_new_counterparty:
          type: boolean
        id:
          type: string
        mandate_required:
          type: boolean
        missing_evidence_action:
          $ref: '#/components/schemas/PolicyAction'
        monthly_minor:
          type:
          - integer
          - 'null'
          format: int64
        on_breach:
          $ref: '#/components/schemas/PolicyAction'
        per_transaction_minor:
          type:
          - integer
          - 'null'
          format: int64
        refund_original_method_only:
          type: boolean
        required_preconditions:
          type: array
          items:
            $ref: '#/components/schemas/FinancialActionPrecondition'
        severity:
          $ref: '#/components/schemas/Severity'
        when:
          $ref: '#/components/schemas/FinancialPolicySelector'
    FinancialPolicySelector:
      type: object
      properties:
        action_kinds:
          type: array
          items:
            $ref: '#/components/schemas/FinancialActionKind'
        agents:
          type: array
          items:
            type: string
        currencies:
          type: array
          items:
            type: string
        operations:
          type: array
          items:
            type: string
        rails:
          type: array
          items:
            $ref: '#/components/schemas/FinancialRail'
    FinancialRail:
      type: string
      enum:
      - payment_http
      - card
      - ach
      - wire
      - internal
      - other
...
    PolicyAction:
      type: string
      description: Action a policy takes when matched.
      enum:
      - block
      - rewrite
      - escalate
    PolicyBatchSetEnabledRequest:
      type: object
      required:
      - ids
      - enabled
      properties:
        enabled:
          type: boolean
        ids:
          type: array
          items:
            type: string
    PolicyBatchSetEnabledResponse:
      type: object
      required:
      - policies
      properties:
        policies:
          type: array
          items:
            $ref: '#/components/schemas/PolicySummary'
    PolicyDocument:
      type: object
      required:
      - id
      - severity
      - enabled
      - source_yaml
      properties:
        description:
          type:
          - string
          - 'null'
        enabled:
          type: boolean
        family:
          $ref: '#/components/schemas/PolicyFamily'
        id:
          type: string
        severity:
          $ref: '#/components/schemas/Severity'
        source_yaml:
          type: string
...
    PolicyFamily:
      type: string
      description: |-
        Stable family discriminator for a policy document.

        `Content` is the legacy/default family used when a policy document does not
        contain a top-level `family:` tag. The other variants are typed policy
        families with their own evaluators and product forms, but they share the
        same policy registry, versioning, and environment deployment lifecycle.
      enum:
      - content
      - flow
      - parameter_source
      - approval
      - memory
      - financial
      - source_label
    PolicyListResponse:
      type: object
      required:
      - policies
      properties:
        policies:
          type: array
          items:
            $ref: '#/components/schemas/PolicySummary'
...
    PolicySetEnabledRequest:
      type: object
      required:
      - enabled
      properties:
        enabled:
          type: boolean
    PolicySummary:
      type: object
      required:
      - id
      - severity
      - enabled
      properties:
        action:
          type:
          - string
          - 'null'
        description:
          type:
          - string
          - 'null'
        enabled:
          type: boolean
        family:
          $ref: '#/components/schemas/PolicyFamily'
        id:
          type: string
        owner_agent_id:
          type:
          - string
          - 'null'
        severity:
          $ref: '#/components/schemas/Severity'
    PolicyValidateResponse:
      type: object
      required:
      - valid
      - errors
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/PolicyValidationIssue'
        policy_id:
          type:
          - string
          - 'null'
        valid:
          type: boolean
    PolicyValidationIssue:
      type: object
      required:
      - path
      - message
      properties:
        message:
          type: string
        path:
          type: string
...
    Severity:
      type: string
      enum:
      - low
      - medium
      - high
      - critical
...